Add back increasing ring feature (2/3).
PS ezio: disabled the ringtone preview when changing the
start volume: it plays at the max volume set
for the ringtone with the main ringtone volume slider, so
it's useless and confusing.
The best would be to play a ringtone preview at the
start volume choosen by the user.

PS: added back the start volume preview (fixed by CM)

patchset: licensing

Change-Id: I8defe1a0dff3df728138c6b43798e722a971d04a